Moved card classes to new subfolder, compiled
