+++ assets/js/script.js
@@ -1,4 +1,4 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const Hand_1 = require("./Hand");
+const Hand_1 = require("./cards/Hand");
 let h = new Hand_1.default();
